:sparkles: solved day 1

+++ AdventOfCode2024/Puzzles/IPuzzle.cs
@@ -2,6 +2,6 @@
 
 public interface IPuzzle
 {
-    int PuzzleOne(Input input);
-    int PuzzleTwo(Input input);
+    long PuzzleOne(Input input);
+    long PuzzleTwo(Input input);
 }
