+++ AdventOfCode2024/Puzzles/PuzzleDay0.cs
@@ -2,12 +2,12 @@
 
 public class PuzzleDay0 : IPuzzle
 {
-    public int PuzzleOne(Input input)
+    public long PuzzleOne(Input input)
     {
         return input.Lines.Select(int.Parse).Sum();
     }
 
-    public int PuzzleTwo(Input input)
+    public long PuzzleTwo(Input input)
     {
         return input.Lines.Select(int.Parse).Aggregate((a, b) => a * b);
     }
